Add a preference to use the latest version of pcap2har.py .
When WebpageTest wants to update pcap2har.py, we do so in two phases.  In the first phase, we check it in to an experimental path, and only those agents that ask to use the experimental version use it.  In the second phase, we move the experimental version to stable, so all clients use it.  This change allows the android mobitest agent to opt in to the experimental version of pcap2har.py .

git-svn-id: http://mobitest-agent.googlecode.com/svn/trunk@7 3cdef708-c80d-ef30-fd85-60ff933634de